+++ packages/anywidget/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @myst-theme/anywidget
 
+## 1.3.0
+
+### Patch Changes
+
+- Updated dependencies [df76193]
+  - @myst-theme/providers@1.3.0
+
+## 1.2.2
+
+### Patch Changes
+
+- @myst-theme/providers@1.2.2
+
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [0888410]
+  - @myst-theme/providers@1.2.1
+
+## 1.2.0
+
+### Patch Changes
+
+- d781f9c: Adding the `myst-anywidget` class name to align with our default class tagging conventions
+- d781f9c: Applies a relative position to the `widgetRoot` which will ensure containment of any absolutely position children
+- Updated dependencies [f908854]
+  - @myst-theme/providers@1.2.0
+
 ## 1.1.4
 
 ### Patch Changes