+++ packages/search-minisearch/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @myst-theme/search-minisearch
 
+## 1.3.0
+
+### Patch Changes
+
+- @myst-theme/search@1.3.0
+
+## 1.2.2
+
+### Patch Changes
+
+- @myst-theme/search@1.2.2
+
+## 1.2.1
+
+### Patch Changes
+
+- @myst-theme/search@1.2.1
+
+## 1.2.0
+
+### Patch Changes
+
+- @myst-theme/search@1.2.0
+
 ## 1.1.4
 
 ### Patch Changes